refactored clickhouse a bit, counts render now
