+++ ui/app/routes.ts
@@ -11,6 +11,10 @@ export default [
     "routes/api/curated_inferences/count.route.ts",
   ),
   route(
+    "api/datasets/count_inserts",
+    "routes/api/datasets/count_inserts.route.ts",
+  ),
+  route(
     "api/function/:function_name/feedback_counts",
     "routes/api/function/$function_name/feedback_counts.route.ts",
   ),
