+++ ui/app/utils/clickhouse/common.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { checkClickHouseConnection } from "./common";
+import { checkClickHouseConnection } from "./client.server";
 
 test("checkClickHouseConnection", async () => {
   const result = await checkClickHouseConnection();
