+++ ui/app/utils/clickhouse/feedback.ts
@@ -1,9 +1,10 @@
-import { z } from "zod";
 import { type TableBounds, TableBoundsSchema } from "./common";
 import { data } from "react-router";
-import { clickhouseClient, InferenceJoinKey } from "./common";
+import { InferenceJoinKey } from "./common";
+import { clickhouseClient } from "./client.server";
 import type { MetricConfig } from "~/utils/config/metric";
 import { getInferenceJoinKey } from "~/utils/clickhouse/curation";
+import { z } from "zod";
 
 export const booleanMetricFeedbackRowSchema = z.object({
   type: z.literal("boolean"),
